Stop truncating retry jitter to whole seconds
calculate_retry_delay added 0-10% jitter and then ran min(int(delay),
max_delay), which quantized the jitter straight back out again. At
error_count=5 the jitter spreads a batch over [16.0, 17.6); truncated, all of
it collapsed onto 16 or 17 -- two buckets, which is the thundering herd the
jitter exists to prevent. run_after is TIMESTAMPTZ with microsecond
resolution, so there was nothing to round for.

Found because it also made the test flaky: test_jitter_prevents_exact_values
draws ten samples and asserts they are not all identical, and ten draws landing
on the same integer second has probability (1/1.6)^10 ~ 0.9%. A full-suite run
hit it. Now it asserts every sample is distinct, which is what a continuous
range actually gives.

Two other tests asserted a symmetric window around the base delay (9.5..10.5
for a base of 10) and passed only because truncation floored every sample back
onto the base. Jitter is one-sided upward, so the real contract is
base <= delay <= base + min(base * 0.1, 5); assert_jittered says it once and
both call sites use it. 40 consecutive runs of the file pass.

+++ pyjobby/retry_strategies.py
@@ -93,8 +93,14 @@ def calculate_retry_delay(
     jitter = random.uniform(0, min(delay * 0.1, 5))
     delay = delay + jitter
 
-    # Cap at max_delay
-    delay = min(int(delay), max_delay)
+    # Cap at max_delay. Deliberately NOT int(): truncating to whole seconds
+    # quantizes the jitter away, which is the one thing jitter must not do.
+    # A batch failing together at error_count=5 spreads over [16.0, 17.6);
+    # truncated, all of it collapses onto 16 or 17 -- two buckets, i.e. the
+    # thundering herd the jitter was added to prevent. run_after is
+    # TIMESTAMPTZ with microsecond resolution, so there is nothing to round
+    # for.
+    delay = min(delay, float(max_delay))
 
     return datetime.timedelta(seconds=delay)
